requirements.txt
# ATD システム依存パッケージ

# Web フレームワーク
Flask==3.0.0
Werkzeug==3.0.1
gunicorn==21.2.0

# データベース
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5

# 認証
Flask-Login==0.6.3
Werkzeug==3.0.1  # パスワードハッシュ化に使用

# その他
python-dotenv==1.0.0  # 環境変数管理